Fixing unsigned integers that could overflow

+++ src/solverconf.h
@@ -238,8 +238,8 @@ class DLL_PUBLIC SolverConf
         long long empty_varelim_time_limitM;
         long long varelim_time_limitM;
         int      updateVarElimComplexityOTF;
-        unsigned updateVarElimComplexityOTF_limitvars;
-        unsigned updateVarElimComplexityOTF_limitavg;
+        uint64_t updateVarElimComplexityOTF_limitvars;
+        uint64_t updateVarElimComplexityOTF_limitavg;
         ElimStrategy  var_elim_strategy; ///<Guess varelim order, or calculate?
         int      varElimCostEstimateStrategy;
         double    varElimRatioPerIter;
